OCR check if entered text can be found

+++ tests/name-game.sikuli/name-game.py
@@ -21,25 +21,30 @@ def main():
         type(Key.BACKSPACE) # hit backspace until name-field is empty
     type(leaderName)
     wait(10)
-    click("ctp2launch-btn.png")
-    if exists("ctp2ctr-bar.png", 100):
-        if waitVanish("ctp2progress-bar.png", 100): # control bar appears before progressbar vanishes
-            if exists(Pattern("ctp2leemur.png").similar(0.90), 100):
-                file = capture(SCREEN.getBounds())
-                if file:
-                    f= Finder(file) # http://doc.sikuli.org/finder.html
-                    f.find("ctp2leemur.png")
-                    if not f.hasNext():
-                        print("Pattern not found in screen shot: " + file)
-                        exit(30)
-                    shutil.move(file, bsfn + '.png')
-                    exit(0)
+    if existsText(leaderName): # OCR check https://sikulix-2014.readthedocs.io/en/latest/news.html#revision-of-the-text-find-api
+        click("ctp2launch-btn.png")
+        if exists("ctp2ctr-bar.png", 100):
+            if waitVanish("ctp2progress-bar.png", 100): # control bar appears before progressbar vanishes
+                if exists(Pattern("ctp2leemur.png").similar(0.90), 100):
+                    file = capture(SCREEN.getBounds())
+                    if file:
+                        f= Finder(file) # http://doc.sikuli.org/finder.html
+                        f.find("ctp2leemur.png")
+                        if not f.hasNext():
+                            print("Pattern not found in screen shot: " + file)
+                            exit(30)
+                        shutil.move(file, bsfn + '.png')
+                        exit(0)
+                else:
+                    exit(35)
             else:
-                exit(35)
+                exit(20)
         else:
-            exit(20)
+            exit(10)
     else:
-        exit(10)
+        file = capture(SCREEN.getBounds())
+        shutil.move(file, bsfn + '.png')
+        exit(40)
     exit(99)
 
 if __name__ == "__main__":
